pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-injection-detector"
version = "1.0.0"
description = "OpenEnv environment for prompt injection detection across three difficulty levels using real HuggingFace security datasets"
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "datasets>=2.18.0",
    "huggingface-hub>=0.21.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[project.scripts]
server = "server.app:main"
prompt-injection-detector = "server.app:main"
